Add custom CA bundle support
In OCP, it's possible to set a cluster-wide proxy that operators can use
to communicate with external APIs. In addition to that, users can also
define a custom CA bundle to be used together with the proxy setting.

Currently, we are not using that custom CA bundle and this patch fixes
that. The main benefit of injecting the CA bundle using the
controllers, as opposed to statically mount it using the manifest file,
is that we can watch for changes in the trusted CA bundle store and roll
out a new Deployment with the new CA bundle.